Small refactor - move extra client logic into service layer
Clean up the client layer by only keeping the web client calls. Move
all additional logic into the service layer.

Update integration tests to assert new layout functions as expected.
